Reordered the GunBase BP and realized the issue with Fire Rate
The Timer being used for the Fire Rate is not checked or updated constantly, only when the gun stops firing, but firing the gun resets the timer, therefore you can shoot with an independent of the set value fire rate if you tap the key instead of holding it.

Solution is to either rework the gun a bit or make a new one entirely.