Reworked activatable system.
Made it a bit modular (in my eyes) to work with and made it work with Slots and Buttons a slight bit better.

However, two issues: buttons are not really compatible with the foam gun (this was a problem before, but overlooked), and getting an error (exception) when trying to access null/None but the error is thrown during the start of a function call for some reason.